Format code

+++ timm/optim/radam.py
@@ -8,8 +8,12 @@ from torch.optim.optimizer import Optimizer, required
 
 
 class RAdam(Optimizer):
-
-    def __init__(self, params, lr=1e-3, betas=(0.9, 0.999), eps=1e-8, weight_decay=0):
+    def __init__(self,
+                 params,
+                 lr=1e-3,
+                 betas=(0.9, 0.999),
+                 eps=1e-8,
+                 weight_decay=0):
         defaults = dict(lr=lr, betas=betas, eps=eps, weight_decay=weight_decay)
         self.buffer = [[None, None, None] for ind in range(10)]
         super(RAdam, self).__init__(params, defaults)
@@ -30,7 +34,8 @@ class RAdam(Optimizer):
                     continue
                 grad = p.grad.data.float()
                 if grad.is_sparse:
-                    raise RuntimeError('RAdam does not support sparse gradients')
+                    raise RuntimeError(
+                        'RAdam does not support sparse gradients')
 
                 p_data_fp32 = p.data.float()
 
@@ -42,7 +47,8 @@ class RAdam(Optimizer):
                     state['exp_avg_sq'] = torch.zeros_like(p_data_fp32)
                 else:
                     state['exp_avg'] = state['exp_avg'].type_as(p_data_fp32)
-                    state['exp_avg_sq'] = state['exp_avg_sq'].type_as(p_data_fp32)
+                    state['exp_avg_sq'] = state['exp_avg_sq'].type_as(
+                        p_data_fp32)
 
                 exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
                 beta1, beta2 = group['betas']
@@ -56,22 +62,25 @@ class RAdam(Optimizer):
                     N_sma, step_size = buffered[1], buffered[2]
                 else:
                     buffered[0] = state['step']
-                    beta2_t = beta2 ** state['step']
+                    beta2_t = beta2**state['step']
                     N_sma_max = 2 / (1 - beta2) - 1
-                    N_sma = N_sma_max - 2 * state['step'] * beta2_t / (1 - beta2_t)
+                    N_sma = N_sma_max - 2 * state['step'] * beta2_t / (1 -
+                                                                       beta2_t)
                     buffered[1] = N_sma
 
                     # more conservative since it's an approximated value
                     if N_sma >= 5:
                         step_size = group['lr'] * math.sqrt(
-                            (1 - beta2_t) * (N_sma - 4) / (N_sma_max - 4) * (N_sma - 2) / N_sma * N_sma_max / (
-                                        N_sma_max - 2)) / (1 - beta1 ** state['step'])
+                            (1 - beta2_t) * (N_sma - 4) / (N_sma_max - 4) *
+                            (N_sma - 2) / N_sma * N_sma_max /
+                            (N_sma_max - 2)) / (1 - beta1**state['step'])
                     else:
-                        step_size = group['lr'] / (1 - beta1 ** state['step'])
+                        step_size = group['lr'] / (1 - beta1**state['step'])
                     buffered[2] = step_size
 
                 if group['weight_decay'] != 0:
-                    p_data_fp32.add_(-group['weight_decay'] * group['lr'], p_data_fp32)
+                    p_data_fp32.add_(-group['weight_decay'] * group['lr'],
+                                     p_data_fp32)
 
                 # more conservative since it's an approximated value
                 if N_sma >= 5:
@@ -86,8 +95,12 @@ class RAdam(Optimizer):
 
 
 class PlainRAdam(Optimizer):
-
-    def __init__(self, params, lr=1e-3, betas=(0.9, 0.999), eps=1e-8, weight_decay=0):
+    def __init__(self,
+                 params,
+                 lr=1e-3,
+                 betas=(0.9, 0.999),
+                 eps=1e-8,
+                 weight_decay=0):
         defaults = dict(lr=lr, betas=betas, eps=eps, weight_decay=weight_decay)
 
         super(PlainRAdam, self).__init__(params, defaults)
@@ -108,7 +121,8 @@ class PlainRAdam(Optimizer):
                     continue
                 grad = p.grad.data.float()
                 if grad.is_sparse:
-                    raise RuntimeError('RAdam does not support sparse gradients')
+                    raise RuntimeError(
+                        'RAdam does not support sparse gradients')
 
                 p_data_fp32 = p.data.float()
 
@@ -120,7 +134,8 @@ class PlainRAdam(Optimizer):
                     state['exp_avg_sq'] = torch.zeros_like(p_data_fp32)
                 else:
                     state['exp_avg'] = state['exp_avg'].type_as(p_data_fp32)
-                    state['exp_avg_sq'] = state['exp_avg_sq'].type_as(p_data_fp32)
+                    state['exp_avg_sq'] = state['exp_avg_sq'].type_as(
+                        p_data_fp32)
 
                 exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
                 beta1, beta2 = group['betas']
@@ -129,22 +144,24 @@ class PlainRAdam(Optimizer):
                 exp_avg.mul_(beta1).add_(1 - beta1, grad)
 
                 state['step'] += 1
-                beta2_t = beta2 ** state['step']
+                beta2_t = beta2**state['step']
                 N_sma_max = 2 / (1 - beta2) - 1
                 N_sma = N_sma_max - 2 * state['step'] * beta2_t / (1 - beta2_t)
 
                 if group['weight_decay'] != 0:
-                    p_data_fp32.add_(-group['weight_decay'] * group['lr'], p_data_fp32)
+                    p_data_fp32.add_(-group['weight_decay'] * group['lr'],
+                                     p_data_fp32)
 
                 # more conservative since it's an approximated value
                 if N_sma >= 5:
                     step_size = group['lr'] * math.sqrt(
-                        (1 - beta2_t) * (N_sma - 4) / (N_sma_max - 4) * (N_sma - 2) / N_sma * N_sma_max / (
-                                    N_sma_max - 2)) / (1 - beta1 ** state['step'])
+                        (1 - beta2_t) * (N_sma - 4) / (N_sma_max - 4) *
+                        (N_sma - 2) / N_sma * N_sma_max /
+                        (N_sma_max - 2)) / (1 - beta1**state['step'])
                     denom = exp_avg_sq.sqrt().add_(group['eps'])
                     p_data_fp32.addcdiv_(-step_size, exp_avg, denom)
                 else:
-                    step_size = group['lr'] / (1 - beta1 ** state['step'])
+                    step_size = group['lr'] / (1 - beta1**state['step'])
                     p_data_fp32.add_(-step_size, exp_avg)
 
                 p.data.copy_(p_data_fp32)
